Add --skip-errors option to webpack-dev-server
Intended to be used when NoErrorsPlugin is active in webpack, but can be
used without it. It should work without --hot too, but untested.

Prevents reloadApp() from being called when there are compilation errors
reported; ultimately stopping webpack/hot/{only-,}dev-server from trying
to fetch a nonexisting update (if NoErrorsPlugin is active) and/or
reloading the window.location.