Handle missing `src` attribute on unused last script
This fixes a bug introduced in 3d3c000, in the case where:

- `__resourceQuery` is set, and
- The last script tag on the page has no `src` attribute

Previously the script tag's `src` attribute was accessed lazily only if
`__resourceQuery` was not defined.  The last script on the page may not be the
script that loaded webpack-dev-server.  This can happen, for example, with the
use of an asynchronous loader such as LABjs.