Fix historyApiFallback, again.
The problem earlier with the fix in #627 was that it broke serving files
without extensions (or with `disableDotRule`).

This passed the test because the `disableDotRule: true` was removed; without
this the test will always succeed because without this rule, a request like `/other.html`
will always go directly to the file.