don't mutate stats configuration (#1174)
The webpack configuration `stats` property has the same format as the
`devServer.stats` property, so the same object is frequently reused to
define the same stats output.

Since webpack v3.8.0, the `stats` property is validated more strictly by
webpack.  webpack-dev-server is mutating this `stats` object by adding a
`colors` property, which is an object and doesn't match the webpack
configuration schema, causing a fatal error.

The easy fix is to *not* mutate the `stats` object, so webpack isn't
breaking with a fatal error.  But we may also consider setting a valid
`colors` property.