Do not use `spdy` on Node 10
`spdy` is effectively unmaintained, and as a consequence of an
implementation that extensively relies on Node’s non-public APIs, broken
on Node 10 and above. In those cases, only https will be used for now.
Once express supports Node's built-in HTTP/2 support, migrating over to
that should be the best way to go.

Fixes: https://github.com/webpack/webpack-dev-server/issues/1449
Fixes: https://github.com/nodejs/node/issues/21665