feat(open): modify `devServer.options` to take an object
https://www.npmjs.com/package/open

Allow for 3 options to pass as options.open:
- Boolean: runs without any options and thus uses the default browser
- String: opens with the browser of your choice
- Object: takes a settings object as defined by https://github.com/sindresorhus/open

As we are now doing now allowing for more complex (and possibly confusing) options to be passed in I have added logging aimed at making this more discoverable to a user.

If they pass in an object with neither `app` nor `wait` OR something that is not ( boolean | string | object ) they will get a warning telling them that they passed in an invalid config and a link to the `open` docs.

With that, I have also hard coded the open failure message, as config related info should now be caught earlier