chore: change log type of "Disconnected!" to info
From my experience over the years, this message should not be an error log.
It is always on the top of my console after a page refresh,
and a few rare cases where I let the devserver run over the night.

It is unfortunate to have to turn off all the useful error or
warning messages to remove it (`client: {logging: 'warn'}`).

Perhaps it is correct for it to be an error.
If so, a [ignoreWarnings](https://webpack.js.org/configuration/other-options/#ignorewarnings)
would perhaps be a better option.