fix: don't reserve unused port too early
At the moment, when starting 2 processes at the same time and using `port: 0` or `port auto` in the majority of cases one of the servers will fail because both server will end up being assigned the same port.

The reason for this is that a number of precious milliseconds elapse between the `getFreePort` and `listen` call. The main reason for this is the `initialize` method.

```
../Users/cli-reproductions/node_modules/webpack-dev-server/lib/Server.js:1544
      throw error;
Error: listen EADDRINUSE: address already in use 127.0.0.1:8080
    at Server.setupListenHandle [as _listen2] (net.js:1320:16)
    at listenInCluster (net.js:1368:12)
    at GetAddrInfoReqWrap.doListen [as callback] (net.js:1505:7)
    at GetAddrInfoReqWrap.onlookup [as oncomplete] (dns.js:71:8) {
  code: 'EADDRINUSE',
  errno: -48,
  syscall: 'listen',
  address: '127.0.0.1',
  port: 8080
}
```