chore: drop strip-ansi dependency
I find copying the strip-ansi source code inside the project
simplifies the set up.

I have noticed that the strip-ansi dependency is precompiled
in production (probably because it is an ES mmodule) and a different
version of strip-ansi is required in the tests. But strip-ansi
is a single function that depends on the ansi-regex package, which exports
a single RegEx. Copying this code inside the project gets rid of two dependencies
and removes the need for precompiling the ansi-regex package.